SIH_Project/requirements.txt
Flask
SQLAlchemy
psycopg2-binary
python-dotenv
werkzeug
pandas
numpy 
scikit-learn
xgboost
